+++ core/src/net/oschina/j2cache/ClusterPolicyFactory.java
@@ -29,9 +29,9 @@ public class ClusterPolicyFactory {
 
     /**
      * 初始化集群消息通知机制
-     * @param broadcast
-     * @param props
-     * @return
+     * @param broadcast  j2cache.broadcast value
+     * @param props  broadcast configuations
+     * @return ClusterPolicy instance
      */
     public final static ClusterPolicy init(String broadcast, Properties props) {
         ClusterPolicy policy;
@@ -39,6 +39,8 @@ public class ClusterPolicyFactory {
             policy = ClusterPolicyFactory.redis(props);
         else if ("jgroups".equalsIgnoreCase(broadcast))
             policy = ClusterPolicyFactory.jgroups(props);//
+        else if ("none".equalsIgnoreCase(broadcast))
+            policy = new NoneClusterPolicy();
         else
             policy = ClusterPolicyFactory.custom(broadcast, props);
         return policy;