+++ core/src/net/oschina/j2cache/redis/RedisUtils.java
@@ -8,14 +8,15 @@ public class RedisUtils {
 
     /**
      * 初始化 Redis 连接池
-     * @param props
-     * @return
+     * @param props j2cache.properties
+     * @param prefix configuration prefix
+     * @return redis connection pool configuration object
      */
     public final static JedisPoolConfig newPoolConfig(Properties props, String prefix) {
         JedisPoolConfig cfg = new JedisPoolConfig();
         cfg.setMaxTotal(Integer.valueOf((String)props.getOrDefault(key(prefix,"maxTotal"), "-1")));
         cfg.setMaxIdle(Integer.valueOf((String)props.getOrDefault(key(prefix,"maxIdle"), "100")));
-        cfg.setMaxWaitMillis(Integer.valueOf((String)props.getOrDefault(key(prefix,"maxWaitMillis"), 100)));
+        cfg.setMaxWaitMillis(Integer.valueOf((String)props.getOrDefault(key(prefix,"maxWaitMillis"), "100")));
         cfg.setMinEvictableIdleTimeMillis(Integer.valueOf((String)props.getOrDefault(key(prefix,"minEvictableIdleTimeMillis"), "864000000")));
         cfg.setMinIdle(Integer.valueOf((String)props.getOrDefault(key(prefix,"minIdle"), "10")));
         cfg.setNumTestsPerEvictionRun(Integer.valueOf((String)props.getOrDefault(key(prefix,"numTestsPerEvictionRun"), "10")));