+++ core/src/net/oschina/j2cache/util/KryoSerializer.java
@@ -30,30 +30,27 @@ import java.io.IOException;
  */
 public class KryoSerializer implements Serializer {
 
-	private final static Kryo kryo = new Kryo();
-
-
     @Override
 	public String name() {
 		return "kryo";
 	}
 
 	@Override
-	public byte[] serialize(Object obj) throws IOException {
+	public byte[] serialize(Object obj) {
 		ByteArrayOutputStream baos = new ByteArrayOutputStream();
 		try (Output output = new Output(baos);){
-			kryo.writeClassAndObject(output, obj);
+			new Kryo().writeClassAndObject(output, obj);
 			output.flush();
 			return baos.toByteArray();
 		}
 	}
 
 	@Override
-	public Object deserialize(byte[] bits) throws IOException {
+	public Object deserialize(byte[] bits) {
 		if(bits == null || bits.length == 0)
 			return null;
 		try (Input ois = new Input(new ByteArrayInputStream(bits))){
-			return kryo.readClassAndObject(ois);
+			return new Kryo().readClassAndObject(ois);
 		}
 	}